Finding unlucky salary persons
"Given tax brackets and given employees gross salaries, determine those employees who
actually get less take-home salary than others with lower initial income."
Suppose the tax deduction law states that:
- No tax will be deducted for persons with salaries ranging from Rs. 0 to Rs. 5,000 per
month or in other words tax deduction rate is 0%.
- 5% tax deduction will be made from the persons with salaries ranging from Rs. 5,001
to Rs. 10,000 per month.
- For persons with salaries ranging from Rs. 10,001 to Rs. 20,000, a 10% tax deduction
rate would be employed.
- For persons with salaries ranging from Rs. 20,001 and higher, 15% tax deduction
would be made.
Taking these rules, let's formulate the problem.
Consider the example of a person with a salary of Rs. 10,000 per month. As per rules,
he/she would be charged by 5% of tax rate. 5% of 10,000 is 500 rupees. So the take home
salary of the person is Rs. 9500.
Now the unfortunate individual, whose gross salary is Rs, 10,001 falls in the next bracket
of tax rate of 10%. He will have to pay tax worth Rs 1000.1. That means the take home
salary of this person is Rs. 9000.9, which is lesser than the person with lower gross salary
of Rs. 10,000. This is the problem.
We can calculate the net salaries of all individuals, determining all the unlucky ones